Added files via upload
Improvements from version 7.

The way this currently works is that after going through the quiz and adding a diary entry, the journal page will show the most recent entry only (previously it would always show the first/oldest entry no matter how many entries you made). Still need to implement a table that displays maybe 10 or so of the most recent entries.

I haven't really bug tested this but the functionality is there. I think I need to add some error handling for if you try to open the journal when the database is empty but that should be an easy thing to fix.